refactor: extract Reporting.ReportGenerator
