+++ .iex.exs
@@ -33,7 +33,15 @@ import Ecto.Query, warn: false
 alias RadiatorWeb.Router.Helpers, as: Routes
 
 alias Radiator.{Directory, Repo, Storage}
-alias Radiator.Directory.{Audio, Episode, Podcast, Network, Editor}
+
+alias Radiator.Directory.{
+  Network,
+  Podcast,
+  Episode,
+  AudioPublication,
+  Audio,
+  Editor
+}
 
 alias Radiator.Media
 alias Radiator.Media.AudioFile
@@ -51,7 +59,11 @@ alias Radiator.Contribution.Person
 alias Radiator.Perm.Ecto.PermissionType
 alias Radiator.Perm.Permission
 
-alias Radiator.Reporting.Report
+alias Radiator.Reporting.{
+  Report,
+  ReportGenerator,
+  ReportWorker
+}
 
 defmodule H do
   def create_episode_audio(episode) do
